chore(admin): 데모 로그인 계정을 prod 프로파일에서도 시드
DemoAccountSeedRunner를 분리해 OPERATOR 2계정(activeventure/hanacard)과
SUPER_ADMIN 데모 계정, 워크스페이스 1·2를 local/dev/prod 에서 멱등 시드한다.
app.workspace/app_user/workspace_member 만 다뤄 domain_pack_version 의존이
없으므로 prod 프로파일에서도 안전하다. 도메인팩 본문 시드는
ActiveVentureDomainPackSeedRunner(local/dev)가 계속 담당한다.

SuperAdminBootstrapRunner 에 @Order 를 부여해 env/secret 기반 운영 SUPER_ADMIN 이
먼저 생성된 뒤 데모 SUPER_ADMIN 이 추가되도록 해 둘이 공존하게 한다(데모가 먼저
들어가면 existsByRole 가드에 운영 계정이 누락됨).

demo-credentials-consistency 테스트의 시드 러너 경로를 DemoAccountSeedRunner 로
갱신하고, 프로파일 게이팅 검증 테스트를 추가한다.
