+++ backend/src/main/java/com/init/domainpack/infrastructure/ActiveVentureDomainPackSeedRunner.java
@@ -46,7 +46,6 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.core.io.ClassPathResource;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -58,9 +57,6 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
   private static final Logger log =
       LoggerFactory.getLogger(ActiveVentureDomainPackSeedRunner.class);
   private static final String DESCRIPTION_FIELD = "description";
-  private static final String DEMO_SIGN_IN_VALUE = String.join("", "demo", "1234");
-  private static final String SUPER_ADMIN_DEMO_EMAIL = "superadmin.demo@ostone.local";
-  private static final String SUPER_ADMIN_DEMO_NAME = "슈퍼 어드민 데모 계정";
   private static final List<SeedConfig> SEED_CONFIGS =
       List.of(
           new SeedConfig(
@@ -87,10 +83,6 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
                   "card_loan_and_cash_advance_management_flow",
                   "card_limit_management_flow",
                   "digital_payment_service_registration_flow")));
-  private static final List<DemoAccountConfig> DEMO_ACCOUNT_CONFIGS =
-      List.of(
-          new DemoAccountConfig(1L, "activeventure.demo@ostone.local", "액티벤처 데모 사용자"),
-          new DemoAccountConfig(2L, "hanacard.demo@ostone.local", "하나카드 데모 사용자"));
 
   private final DomainPackCommandRepository domainPackRepository;
   private final DomainPackVersionRepository domainPackVersionRepository;
@@ -103,7 +95,6 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
   private final WorkflowMatchingProfileBuildRequestService profileBuildRequestService;
   private final EntityManager entityManager;
   private final ObjectMapper objectMapper;
-  private final PasswordEncoder passwordEncoder;
 
   public ActiveVentureDomainPackSeedRunner(
       DomainPackCommandRepository domainPackRepository,
@@ -116,8 +107,7 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
       IntentSlotBindingRepository intentSlotBindingRepository,
       WorkflowMatchingProfileBuildRequestService profileBuildRequestService,
       EntityManager entityManager,
-      ObjectMapper objectMapper,
-      PasswordEncoder passwordEncoder) {
+      ObjectMapper objectMapper) {
     this.domainPackRepository = domainPackRepository;
     this.domainPackVersionRepository = domainPackVersionRepository;
     this.intentDefinitionRepository = intentDefinitionRepository;
@@ -129,7 +119,6 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
     this.profileBuildRequestService = profileBuildRequestService;
     this.entityManager = entityManager;
     this.objectMapper = objectMapper;
-    this.passwordEncoder = passwordEncoder;
   }
 
   @Override
@@ -138,7 +127,6 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
     for (SeedConfig seedConfig : SEED_CONFIGS) {
       seed(seedConfig);
     }
-    seedDemoAccounts();
   }
 
   private void seed(SeedConfig seedConfig) {
@@ -232,124 +220,6 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
     entityManager
         .createNativeQuery(
             "SELECT setval('app.workspace_id_seq', (SELECT COALESCE(MAX(id), 1) FROM app.workspace), true)")
-        .getSingleResult();
-  }
-
-  private void seedDemoAccounts() {
-    for (DemoAccountConfig accountConfig : DEMO_ACCOUNT_CONFIGS) {
-      Long userId = upsertDemoUser(accountConfig);
-      upsertDemoMembership(accountConfig.workspaceId(), userId);
-      log.info(
-          "Seed demo account '{}' mapped to workspace {}",
-          accountConfig.email(),
-          accountConfig.workspaceId());
-    }
-    seedSuperAdminAccount();
-    resetAppUserSequence();
-    resetWorkspaceMemberSequence();
-  }
-
-  // 워크스페이스에 종속되지 않는 글로벌 SUPER_ADMIN 데모 계정을 멱등 upsert 한다.
-  // SUPER_ADMIN 전용 콘솔/관리 API 로컬 시연·검증용이며, 운영자 데모 계정과 동일한
-  // DEMO_SIGN_IN_VALUE 비밀번호를 사용한다(workspace_member 매핑은 만들지 않는다).
-  private void seedSuperAdminAccount() {
-    entityManager
-        .createNativeQuery(
-            """
-            INSERT INTO app.app_user (
-              email, name, password_hash, password_reset_required, role, status, profile_json
-            )
-            VALUES (
-              :email, :name, :credentialHash, false, 'SUPER_ADMIN', 'ACTIVE', '{}'::jsonb
-            )
-            ON CONFLICT (email) DO UPDATE
-              SET name = EXCLUDED.name,
-                  password_hash = EXCLUDED.password_hash,
-                  password_reset_required = false,
-                  role = 'SUPER_ADMIN',
-                  status = 'ACTIVE',
-                  profile_json = '{}'::jsonb,
-                  password_reset_token_hash = null,
-                  password_reset_token_expires_at = null,
-                  updated_at = now()
-            """)
-        .setParameter("email", SUPER_ADMIN_DEMO_EMAIL)
-        .setParameter("name", SUPER_ADMIN_DEMO_NAME)
-        .setParameter("credentialHash", passwordEncoder.encode(DEMO_SIGN_IN_VALUE))
-        .executeUpdate();
-    log.info("Seed super admin demo account '{}'", SUPER_ADMIN_DEMO_EMAIL);
-  }
-
-  private Long upsertDemoUser(DemoAccountConfig accountConfig) {
-    Object result =
-        entityManager
-            .createNativeQuery(
-                """
-                INSERT INTO app.app_user (
-                  email, name, password_hash, password_reset_required, role, status, profile_json
-                )
-                VALUES (
-                  :email, :name, :credentialHash, false, 'OPERATOR', 'ACTIVE', '{}'::jsonb
-                )
-                ON CONFLICT (email) DO UPDATE
-                  SET name = EXCLUDED.name,
-                      password_hash = EXCLUDED.password_hash,
-                      password_reset_required = false,
-                      role = 'OPERATOR',
-                      status = 'ACTIVE',
-                      profile_json = '{}'::jsonb,
-                      password_reset_token_hash = null,
-                      password_reset_token_expires_at = null,
-                      updated_at = now()
-                RETURNING id
-            """)
-            .setParameter("email", accountConfig.email())
-            .setParameter("name", accountConfig.name())
-            .setParameter("credentialHash", passwordEncoder.encode(DEMO_SIGN_IN_VALUE))
-            .getSingleResult();
-    if (result instanceof Number number) {
-      return number.longValue();
-    }
-    throw new IllegalStateException("Demo account upsert did not return a numeric user id");
-  }
-
-  private void upsertDemoMembership(Long workspaceId, Long userId) {
-    entityManager
-        .createNativeQuery(
-            """
-            INSERT INTO app.workspace_member (workspace_id, user_id, member_role)
-            VALUES (:workspaceId, :userId, 'OPERATOR')
-            ON CONFLICT (workspace_id, user_id) DO UPDATE
-              SET member_role = EXCLUDED.member_role
-            """)
-        .setParameter("workspaceId", workspaceId)
-        .setParameter("userId", userId)
-        .executeUpdate();
-  }
-
-  private void resetAppUserSequence() {
-    entityManager
-        .createNativeQuery(
-            """
-            SELECT setval(
-              'app.app_user_id_seq',
-              (SELECT COALESCE(MAX(id), 1) FROM app.app_user),
-              true
-            )
-            """)
-        .getSingleResult();
-  }
-
-  private void resetWorkspaceMemberSequence() {
-    entityManager
-        .createNativeQuery(
-            """
-            SELECT setval(
-              'app.workspace_member_id_seq',
-              (SELECT COALESCE(MAX(id), 1) FROM app.workspace_member),
-              true
-            )
-            """)
         .getSingleResult();
   }
 
@@ -1055,6 +925,4 @@ public class ActiveVentureDomainPackSeedRunner implements ApplicationRunner {
       String description,
       String profileBuildSource,
       Set<String> policyHandoffWorkflowCodes) {}
-
-  private record DemoAccountConfig(Long workspaceId, String email, String name) {}
 }
